Renames v2 transformer file
